my-cli-test-app/services/generated/microservice/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
pydantic==2.4.2
numpy==1.24.3
cirq-core==1.2.0
qiskit==0.44.1
amazon-braket-sdk==1.39.0
